move python files to app/ change file structure in general fix db.registerUser and db.getLastPosts
